update Readme; fix bugs

+++ function/andor/readAscImage.m
@@ -1,4 +1,4 @@
 function image = readAscImage(path, file)
     image = readmatrix(fullfile(path, file), "FileType","text");
-    image = flip(transpose(image(:, 2:end)));    
+    image = uint16(flip(transpose(image(:, 2:end))));
 end
